+++ Tests/FunctionalTests/Sources/FTRBaseIntegrationTest.h
@@ -14,6 +14,9 @@
 // limitations under the License.
 //
 
+/* Define for checking if the current iOS version is 11 or above. */
+#define iOS11_OR_ABOVE() ([UIDevice currentDevice].systemVersion.intValue >= 11)
+
 // Base test class for all EarlGrey integration tests.
 @interface FTRBaseIntegrationTest : XCTestCase